requirements.txt
Django==3.2
djangorestframework==3.11.0
django-environ==0.8.1
Pillow==9.0.0
